Ep3. Revenge of the Small Dollar Donor.
War! The galactic Republicans are crumbling! The world is in chaos! There is evil everywhere. But there are unlikely heroes abound. 

In the Northern region of the galaxy, on the planet of Virginia, relative peace is being achieved. In this time of calm, state and local Jedi’s have raised over 180,000 for their new campaigns in the last 30 days. 

In the Midwest section of the universe the battle still rages on. In Wisconsin, the Jedi have almost reached 4million dollars for their cause. 

The ActBlue Jedi order must remain strong! The next battle awaits them...